SmartApp Server: Adding iPhone push notification support
- Enabled iPhone push notification capabilities.
- Now robot notifications to smart apps that go as push notifications can handle both
  iOS and Android devices.
- Added web interface to send iPhone notifications to users.
- Minor UI fixes

Note that the iOS plugins with push notifications support have NOT been checked in yet.
We have tested this internally with test versions - once testing is complete, the iOS smartapps.

Important: The server needs the app's Apple certificates in order for push notifications to work.
Currently, we have used Raja Software's Apple certificates for testing - this will NOT work with
UID builds. When we have a final certificates, we need to upload these to the server and change
the code to use the updated certs.